chore: move base grid option to `src\webviews\pages`
